+++ Change Log.md
@@ -30,3 +30,11 @@
    - **Actual Change**:
      1. 给出了便于调试的两个服务，分别使机械臂竖直和放下
      2. 使切割曲线变得完整
+ - **Header**:  76c191
+    - **comments** : update with scale speed function and linear move and fix some bugs
+    - **Actual Change**:
+        1. 修复了参数监视器析构的bug，并且使单线程监听多个参数，节约了资源
+        2. 在相机节点和主节点中添加了看门狗机制以监视话题消失
+        3. 在主函数中添加了两种线性运动，和轨迹速度缩放
+        4. 使tag检测能居中
+        5. 添加了依赖安装脚本